LayoutTests/http/tests/security/xssAuditor/full-block-javascript-link-expected.txt

CONSOLE MESSAGE: line 14: Refused to execute a JavaScript script. Source code of script found within request.

CONSOLE MESSAGE: Entire page will be blocked.
CONSOLE MESSAGE: Unsafe JavaScript attempt to access frame with URL data:text/html,<p></p> from frame with URL http://127.0.0.1:8000/security/xssAuditor/full-block-javascript-link.html. The frame requesting access has a protocol of 'http', the frame being accessed has a protocol of ''. Protocols must match.

ALERT: URL mismatch: undefined vs. http://127.0.0.1:8000/security/xssAuditor/resources/echo-intertag-click-and-notify.pl?enable-full-block=1&elmid=anchorLink&q=%3Ca+id%3DanchorLink+href%3Djavascript%3Aalert%280%29%3Etest%3C/a%3E
There should be no content in the iframe below:



--------
Frame: 'frame'
--------
